fc_db::kv

Enum DatabaseSource

pub enum DatabaseSource {
    Auto {
        paritydb_path: PathBuf,
        rocksdb_path: PathBuf,
        cache_size: usize,
    },
    RocksDb {
        path: PathBuf,
        cache_size: usize,
    },
    ParityDb {
        path: PathBuf,
    },
    Custom {
        db: Arc<dyn Database<H256>>,
        require_create_flag: bool,
    },
}

Where to find the database..

Variants

Auto

Check given path, and see if there is an existing database there. If itโs either RocksDb or ParityDb, use it. If there is none, create a new instance of ParityDb.

Fields

paritydb_path: PathBuf

Path to the paritydb database.

rocksdb_path: PathBuf

Path to the rocksdb database.

cache_size: usize

Cache size in MiB. Used only by RocksDb variant of DatabaseSource.

RocksDb

Load a RocksDB database from a given path. Recommended for most uses.

Fields

path: PathBuf

Path to the database.

cache_size: usize

Cache size in MiB.

ParityDb

Load a ParityDb database from a given path.

Fields

path: PathBuf

Path to the database.

Custom

Use a custom already-open database.

Fields

db: Arc<dyn Database<H256>>

the handle to the custom storage

require_create_flag: bool

if set, the create flag will be required to open such datasource

Implementations

impl DatabaseSource

pub fn path(&self) -> Option<&Path>

Return path for databases that are stored on disk.

pub fn set_path(&mut self, p: &Path) -> bool

Set path for databases that are stored on disk.