fc_db

Module sql

Structs

Backend

BlockIndexedStatus

Represents the indexed status of a block and if it’s canon or not.

Log

Represents a log item.

SqliteBackendConfig

Represents the Sqlite connection options that are used to establish a database connection.

Enums

BackendConfig

Represents the backend configurations.