Struct fc_db::sql::Backend

pub struct Backend<Block: BlockT> { /* private fields */ }

Implementations

impl<Block> Backend<Block>where Block: BlockT<Hash = H256> + BlockT,

pub async fn new( config: BackendConfig<'_>, pool_size: u32, num_ops_timeout: Option<NonZeroU32>, overrides: Arc<OverrideHandle<Block>> ) -> Result<Self, Error>

Creates a new instance of the SQL backend.

pub fn pool(&self) -> &SqlitePool

Get the underlying Sqlite pool.

pub async fn canonicalize( &self, retracted: &[H256], enacted: &[H256] ) -> Result<(), Error>

Canonicalize the indexed blocks, marking/demarking them as canon based on the provided retracted and enacted values.

pub async fn insert_genesis_block_metadata<Client, BE>( &self, client: Arc<Client> ) -> Result<Option<H256>, Error>where Client: StorageProvider<Block, BE> + HeaderBackend<Block> + 'static + ProvideRuntimeApi<Block>, Client::Api: EthereumRuntimeRPCApi<Block>, BE: BackendT<Block> + 'static, BE::State: StateBackend<BlakeTwo256>,

Index the block metadata for the genesis block.

pub async fn insert_block_metadata<Client, BE>( &self, client: Arc<Client>, hash: H256 ) -> Result<(), Error>where Client: StorageProvider<Block, BE> + HeaderBackend<Block> + 'static, BE: BackendT<Block> + 'static, BE::State: StateBackend<BlakeTwo256>,

Insert the block metadata for the provided block hashes.

pub async fn index_block_logs<Client, BE>( &self, client: Arc<Client>, block_hash: Block::Hash )where Client: StorageProvider<Block, BE> + HeaderBackend<Block> + 'static, BE: BackendT<Block> + 'static, BE::State: StateBackend<BlakeTwo256>,

Index the logs for the newly indexed blocks upto a max_pending_blocks value.

pub async fn is_block_indexed(&self, block_hash: Block::Hash) -> bool

Retrieves the status if a block has been already indexed.

pub async fn block_indexed_and_canon_status( &self, block_hash: Block::Hash ) -> BlockIndexedStatus

Retrieves the status if a block is indexed and if also marked as canon.

pub async fn set_block_as_canon( &self, block_hash: H256 ) -> Result<SqliteQueryResult, Error>

Sets the provided block as canon.

pub async fn get_first_missing_canon_block(&self) -> Option<u32>

Retrieves the first missing canonical block number in decreasing order that hasn’t been indexed yet. If no unindexed block exists or the table or the rows do not exist, then the function returns None.

pub async fn get_first_pending_canon_block(&self) -> Option<H256>

Retrieves the first pending canonical block hash in decreasing order that hasn’t had then the function returns None.

pub async fn get_last_indexed_canon_block(&self) -> Result<H256, Error>

Retrieve the block hash for the last indexed canon block.